+++ .eslintignore
@@ -1,4 +1,3 @@
 webpack.config.js
 webpack.client.config.js
 webpack.server.config.js
-/src/server/server.js
