+++ package.json
@@ -7,7 +7,9 @@
   "license": "MIT",
   "dependencies": {
     "@hapi/hapi": "^20.1.3",
+    "@hapi/inert": "^6.0.3",
     "classnames": "^2.3.1",
+    "handlebars": "^4.7.7",
     "hookrouter": "1.2.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
