+++ public/index.html
@@ -8,6 +8,6 @@
     <title>PokemonDesk</title>
   </head>
   <body>
-    <div id="root"></div>
+    <div id="root">{{{content}}}</div>
   </body>
 </html>
