--- src/server/server.js
@@ -1,27 +1,0 @@
-'use strict';
-const Hapi = require('@hapi/hapi');
-
-const init = async () => {
-  const server = Hapi.server({
-    port: 3000,
-    host: 'localhost',
-  });
-
-  server.route({
-    method: 'GET',
-    path: '/{any*}',
-    handler: (request, h) => {
-      return `Hello World! ${request.path}`;
-    },
-  });
-
-  await server.start();
-  console.log('Server running on %s', server.info.uri);
-};
-
-process.on('unhandledRejection', (err) => {
-  console.log(err);
-  process.exit(1);
-});
-
-init();
